+++ src/boostLogging.ts
@@ -28,12 +28,14 @@ export class BoostLogger extends Disposable {
         this.log("INFO: " + message);
         window.showInformationMessage(message);
     }
+
     warn(message: string, showUI : boolean = true) {
         this.log("WARNING: " + message);
         if (showUI) {
             window.showWarningMessage(message);
         }
     }
+    
     error(message: string, showUI : boolean = true) {
         this.log("ERROR: " + message);
         if (showUI) {