+++ src/extension.ts
@@ -6,6 +6,8 @@ import { BoostConvertKernel } from './convert_controller';
 import { BoostComplianceKernel } from './compliance_controller';
 import { BoostExplainKernel, explainCellMarker } from './explain_controller';
 import { BoostCodeGuidelinesKernel } from './codeguidelines_controller';
+import { BoostArchitectureBlueprintKernel } from './blueprint_controller';
+import { BoostCustomProcessKernel, customProcessCellMarker } from './custom_controller';
 
 import { BoostContentSerializer } from './serializer';
 import { parseFunctions } from './split';	
@@ -16,7 +18,6 @@ import { KernelControllerBase} from './base_controller';
 import { TextDecoder, TextEncoder } from 'util';
 import * as fs from 'fs';
 import * as path from 'path';
-import { BoostArchitectureBlueprintKernel } from './blueprint_controller';
 import * as boostnb from './jupyter_notebook';
 import { registerCustomerPortalCommand, setupBoostStatus } from './portal';
 
@@ -181,6 +182,14 @@ function setupNotebookEnvironment(
         guidelinesKernel,
         blueprintKernel
 	);
+  
+        // if in dev mode, register all dev only kernels
+    if (BoostConfiguration.enableDevOnlyKernels) {
+        let customProcessKernel = new BoostCustomProcessKernel(collection);
+        kernelMap.set(customProcessKernel.outputType, customProcessKernel);
+        context.subscriptions.push(customProcessKernel);
+    }    
+
 }
 
 function registerOpenCodeFile(context: vscode.ExtensionContext) {
@@ -544,6 +553,9 @@ function registerFileRightClickAnalyzeCommand(context: vscode.ExtensionContext, 
                 }
 
                 await parseFunctionsFromFile(uri, currentNotebook);
+
+                boostLogging.log(`Boosted file:[${uri.fsPath.toString()}`);
+                vscode.window.showNotebookDocument(currentNotebook);
             } catch (error) {
                 boostLogging.error(`Unable to Boost file:[${uri.fsPath.toString()} due to error:${error}`);
             }