+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Polyverse Boost Automatic Application Modernization
 ### Enhancements
 - Enabled Folder and Project processing to build Notebook files in background as files - avoid opening Notebook files in Editor
 - Improved visual error messages when an error occurs processing an entire project or folder
+- Added analysis of software licenses used in source code to the Architectural Blueprint Summary report
 
 ## Version 0.9.4: April 24, 2023